requirements.txt
fastapi==0.81.0
uvicorn==0.18.3
sqlalchemy
pydantic[dotenv]
psycopg2-binary
alembic
python-multipart
python-jose
passlib
bcrypt
